+++ internal/cmd/company/iam/edit.go
@@ -17,6 +17,7 @@ package iam
 
 import (
 	"github.com/mia-platform/miactl/internal/clioptions"
+	"github.com/mia-platform/miactl/internal/cmd/company/iam/serviceaccount"
 	"github.com/mia-platform/miactl/internal/cmd/company/iam/user"
 	"github.com/spf13/cobra"
 )
@@ -30,6 +31,7 @@ func EditCmd(options *clioptions.CLIOptions) *cobra.Command {
 
 	cmd.AddCommand(
 		user.EditCmd(options),
+		serviceaccount.EditCmd(options),
 	)
 
 	return cmd