+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `company iam edit user` command
 - `company iam add group` command
 - `company iam add group-member` command
+- `company iam edit serviceaccount` command
 
 ### Changed